Fixed 64bit compatibility for bi-directional protocol serialization/de-serialization Cleanup, fixes all the warnings for Bluetooth_Dongle, Windows version.

+++ src/VIPShared/BD_CP_SetMenuImages.cpp
@@ -74,7 +74,7 @@ void BD_CP_SetMenuImages::ConvertCommandToBinary()
 {
 	BDCommand::ConvertCommandToBinary();
 	Write_unsigned_char( m_nItemsType );
-	Write_long( m_vItems.size() );
+	Write_long( (long)m_vItems.size() );
 	for (vector<MenuItemInfo*>::iterator iter=m_vItems.begin(); iter!=m_vItems.end(); ++iter ) {
 		(*iter)->write( m_nItemsType, *this );
 	}
